refactor: features/tags, Header into blocks/Layout, group LedgerEntry blocks
- Tags/TagDimensionsGroup -> features/tags, not @blocks. TagDimensionCombobox
  fetches via useGetTagDimensionByKey, which would have made it the first block to
  import @api. A feature that fetches is ordinary. Cost: its four consumers span
  bankTransactions and generalLedger, so features/tags is cross-domain from day one.
- @ui/Header -> @blocks/Layout/Header, grouping it with the rest of the page
  scaffolding. No @ui or @blocks component imported it, so no layering inversion.
- blocks/Layout/Container/Header.tsx -> blocks/Layout/DeprecatedHeader, exporting
  DeprecatedHeader; its three consumers updated. CSS class left alone, so the two
  Headers are still to be collapsed for real later.
- LedgerEntryDetailField, LedgerEntryDetailSection, LedgerEntrySourceDetailView
  grouped under @blocks/LedgerEntry/, each keeping its own subdirectory. All three
  are used by both generalLedger and profitAndLoss, which is what keeps them blocks.
