+++ src/components/blocks/LedgerEntry/LedgerEntryDetailField/LedgerEntryDetailField.tsx
