+++ src/components/blocks/LedgerEntry/LedgerEntryDetailField/ledgerEntryDetailField.scss
