+++ src/components/blocks/LedgerEntry/LedgerEntryDetailSection/LedgerEntryDetailSection.tsx
