+++ src/components/blocks/LedgerEntry/LedgerEntryDetailSection/ledgerEntryDetailSection.scss
