+++ src/components/blocks/LedgerEntry/LedgerEntrySourceDetailView/LedgerEntrySourceDetailView.tsx
@@ -5,7 +5,7 @@ import { type LedgerEntrySourceType } from '@schemas/generalLedger/ledgerEntrySo
 import { humanizeEnum } from '@utils/format'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { DateTime } from '@ui/DateTime/DateTime'
-import { LedgerEntryDetailField } from '@blocks/LedgerEntryDetailField/LedgerEntryDetailField'
+import { LedgerEntryDetailField } from '@blocks/LedgerEntry/LedgerEntryDetailField/LedgerEntryDetailField'
 
 export interface LedgerEntrySourceDetailStringOverrides {
   sourceLabel?: string
