+++ src/components/features/generalLedger/JournalEntryDetailHeader/JournalEntryDetailHeader.tsx
@@ -2,10 +2,10 @@ import { type ReactNode } from 'react'
 
 import { BackButton } from '@ui/Button/BackButton'
 import { CloseButton } from '@ui/Button/CloseButton'
-import { Header } from '@ui/Header/Header'
-import { HeaderCol } from '@ui/Header/HeaderCol'
-import { HeaderRow } from '@ui/Header/HeaderRow'
 import { Heading } from '@ui/Typography/Heading'
+import { Header } from '@blocks/Layout/Header/Header'
+import { HeaderCol } from '@blocks/Layout/Header/HeaderCol'
+import { HeaderRow } from '@blocks/Layout/Header/HeaderRow'
 
 import './journalEntryDetailHeader.scss'
