+++ src/components/features/generalLedger/JournalEntryForm/JournalEntryForm.tsx
@@ -14,10 +14,10 @@ import { DataState, DataStateStatus } from '@ui/DataState/DataState'
 import { Form } from '@ui/Form/Form'
 import { Separator } from '@ui/Separator/Separator'
 import { HStack, VStack } from '@ui/Stack/Stack'
-import { TagDimensionsGroup } from '@blocks/TagDimensionsGroup/TagDimensionsGroup'
 import { CustomerVendorSelector } from '@features/customerVendor/CustomerVendorSelector/CustomerVendorSelector'
 import { JournalEntryLineItemsTable } from '@features/generalLedger/JournalEntryForm/JournalEntryLineItemsTable'
 import { useJournalEntryForm } from '@features/generalLedger/JournalEntryForm/useJournalEntryForm'
+import { TagDimensionsGroup } from '@features/tags/TagDimensionsGroup/TagDimensionsGroup'
 
 import './journalEntryForm.scss'
