+++ src/components/features/generalLedger/JournalTableWithPanel/JournalTableWithPanel.tsx
@@ -7,11 +7,11 @@ import { useJournalNavigation } from '@providers/JournalStore/JournalStoreProvid
 import { useBookkeepingStatusContext } from '@contexts/BookkeepingStatusContext/BookkeepingStatusContext'
 import { JournalContext } from '@contexts/JournalContext/JournalContext'
 import { Button } from '@ui/Button/Button'
-import { Header } from '@ui/Header/Header'
-import { HeaderCol } from '@ui/Header/HeaderCol'
-import { HeaderRow } from '@ui/Header/HeaderRow'
 import { Heading } from '@ui/Typography/Heading'
 import { LedgerDateRangeSelection } from '@blocks/DatePickers/DateSelection/LedgerDateRangeSelection'
+import { Header } from '@blocks/Layout/Header/Header'
+import { HeaderCol } from '@blocks/Layout/Header/HeaderCol'
+import { HeaderRow } from '@blocks/Layout/Header/HeaderRow'
 import { Panel } from '@blocks/Layout/View/Panel/Panel'
 import { JournalEntriesDownloadButton } from '@features/generalLedger/JournalEntriesDownloadButton/JournalEntriesDownloadButton'
 import { JournalSidebar } from '@features/generalLedger/JournalSidebar/JournalSidebar'
