+++ src/components/features/generalLedger/LedgerEntryDetails/LedgerEntryDetails.tsx
@@ -9,9 +9,9 @@ import { useInAppLinkContext } from '@contexts/InAppLinkContext'
 import { Badge } from '@ui/Badge/Badge'
 import { DateTime } from '@ui/DateTime/DateTime'
 import { VStack } from '@ui/Stack/Stack'
-import { LedgerEntryDetailField } from '@blocks/LedgerEntryDetailField/LedgerEntryDetailField'
-import { LedgerEntryDetailSection } from '@blocks/LedgerEntryDetailSection/LedgerEntryDetailSection'
-import { LedgerEntrySourceDetailView } from '@blocks/LedgerEntrySourceDetailView/LedgerEntrySourceDetailView'
+import { LedgerEntryDetailField } from '@blocks/LedgerEntry/LedgerEntryDetailField/LedgerEntryDetailField'
+import { LedgerEntryDetailSection } from '@blocks/LedgerEntry/LedgerEntryDetailSection/LedgerEntryDetailSection'
+import { LedgerEntrySourceDetailView } from '@blocks/LedgerEntry/LedgerEntrySourceDetailView/LedgerEntrySourceDetailView'
 import { JournalEntryDetailHeader } from '@features/generalLedger/JournalEntryDetailHeader/JournalEntryDetailHeader'
 import { LedgerEntryDetailsLineItemsTable } from '@features/generalLedger/LedgerEntryDetails/LedgerEntryDetailsLineItemsTable'
 import { ReverseLedgerEntryButton } from '@features/generalLedger/ReverseLedgerEntryButton/ReverseLedgerEntryButton'
