+++ src/components/features/profitAndLoss/ProfitAndLossDetailReport/ProfitAndLossDetailReport.tsx
@@ -19,9 +19,9 @@ import { MoneySpan } from '@ui/Typography/MoneySpan'
 import { Label, Span } from '@ui/Typography/Text'
 import type { ColumnConfig } from '@blocks/DataTable/utils/column'
 import { BaseDetailView } from '@blocks/Layout/BaseDetailView/BaseDetailView'
-import { LedgerEntryDetailField } from '@blocks/LedgerEntryDetailField/LedgerEntryDetailField'
-import { LedgerEntryDetailSection } from '@blocks/LedgerEntryDetailSection/LedgerEntryDetailSection'
-import { LedgerEntrySourceDetailView } from '@blocks/LedgerEntrySourceDetailView/LedgerEntrySourceDetailView'
+import { LedgerEntryDetailField } from '@blocks/LedgerEntry/LedgerEntryDetailField/LedgerEntryDetailField'
+import { LedgerEntryDetailSection } from '@blocks/LedgerEntry/LedgerEntryDetailSection/LedgerEntryDetailSection'
+import { LedgerEntrySourceDetailView } from '@blocks/LedgerEntry/LedgerEntrySourceDetailView/LedgerEntrySourceDetailView'
 import { VirtualizedDataTable } from '@blocks/VirtualizedDataTable/VirtualizedDataTable'
 import { type BreadcrumbItem, ProfitAndLossDetailReportBreadcrumb } from '@features/profitAndLoss/ProfitAndLossDetailReportBreadcrumb/ProfitAndLossDetailReportBreadcrumb'
