+++ src/components/features/profitAndLoss/ProfitAndLossReport/ProfitAndLossReport.tsx
@@ -5,11 +5,11 @@ import type { TimeRangePickerConfig } from '@internal-types/reports'
 import { useReportsCompactHeader } from '@hooks/features/reports/useReportsCompactHeader'
 import { InAppLinkProvider, type LinkingMetadata } from '@contexts/InAppLinkContext'
 import { ProfitAndLossContext } from '@contexts/ProfitAndLossContext/ProfitAndLossContext'
-import { Header } from '@ui/Header/Header'
-import { HeaderCol } from '@ui/Header/HeaderCol'
-import { HeaderRow } from '@ui/Header/HeaderRow'
 import { HStack, Stack } from '@ui/Stack/Stack'
 import { CombinedDateRangeSelection } from '@blocks/DatePickers/DateSelection/CombinedDateRangeSelection'
+import { Header } from '@blocks/Layout/Header/Header'
+import { HeaderCol } from '@blocks/Layout/Header/HeaderCol'
+import { HeaderRow } from '@blocks/Layout/Header/HeaderRow'
 import { View } from '@blocks/Layout/View/View'
 import { ProfitAndLossDetailReport } from '@features/profitAndLoss/ProfitAndLossDetailReport/ProfitAndLossDetailReport'
 import { type BreadcrumbItem } from '@features/profitAndLoss/ProfitAndLossDetailReportBreadcrumb/ProfitAndLossDetailReportBreadcrumb'
