+++ src/components/features/reports/StatementOfCashFlow/StatementOfCashFlow.tsx
@@ -8,11 +8,11 @@ import { useReportsCompactHeader } from '@hooks/features/reports/useReportsCompa
 import { useResolvedReportView } from '@hooks/features/reports/useResolvedReportView'
 import { useGlobalDateRange } from '@providers/DateStoreProvider/GlobalDateStoreProvider'
 import { TableProvider } from '@contexts/TableContext/TableContext'
-import { Header } from '@ui/Header/Header'
-import { HeaderCol } from '@ui/Header/HeaderCol'
-import { HeaderRow } from '@ui/Header/HeaderRow'
 import { HStack, Stack } from '@ui/Stack/Stack'
 import { CombinedDateRangeSelection } from '@blocks/DatePickers/DateSelection/CombinedDateRangeSelection'
+import { Header } from '@blocks/Layout/Header/Header'
+import { HeaderCol } from '@blocks/Layout/Header/HeaderCol'
+import { HeaderRow } from '@blocks/Layout/Header/HeaderRow'
 import { View } from '@blocks/Layout/View/View'
 import { ConditionalBlock } from '@components/utility/ConditionalBlock'
 import { CashflowStatementDownloadButton } from '@features/reports/CashflowStatementDownloadButton/CashflowStatementDownloadButton'
